Mapfeatures2

+++ src/main/java/ca/concordia/app/risk/services/MapService.java
@@ -5,6 +5,7 @@ import org.springframework.beans.BeanUtils;
 import ca.concordia.app.risk.controller.dto.BorderDto;
 import ca.concordia.app.risk.controller.dto.ContinentDto;
 import ca.concordia.app.risk.controller.dto.CountryDto;
+import ca.concordia.app.risk.exceptions.RiskGameRuntimeException;
 import ca.concordia.app.risk.model.cache.RunningGame;
 import ca.concordia.app.risk.model.dao.BorderDaoImp;
 import ca.concordia.app.risk.model.dao.ContinentDaoImpl;
@@ -22,10 +23,14 @@ public class MapService {
 	 * 
 	 * @param continentDto
 	 */
-	public void addContinent(ContinentDto continentDto) throws Exception {
-		ContinentModel continentModel = objectFactory.createContinentModel();
-		BeanUtils.copyProperties(continentDto, continentModel);
+	public void addContinent(ContinentDto continentDto) {
 		ContinentDaoImpl continentDaoImp = new ContinentDaoImpl();
+		ContinentModel continentModel = continentDaoImp.findByName(RunningGame.getInstance(), continentDto.getName());
+		if(continentModel != null) {
+			throw new RiskGameRuntimeException(String.format("Continent [%s] already exist, cannot be added!", continentDto.getName())); 
+		}
+		continentModel = objectFactory.createContinentModel();
+		BeanUtils.copyProperties(continentDto, continentModel);
 		continentDaoImp.assignID(RunningGame.getInstance(), continentModel);
 		RunningGame.getInstance().getContinents().getList().add(continentModel);
 	}
@@ -34,18 +39,17 @@ public class MapService {
 	 * 
 	 * @param continentDto
 	 */
-	public void removeContinent(ContinentDto continentDto) throws Exception {
+	public void removeContinent(ContinentDto continentDto){
 		ContinentDaoImpl continentDao = new ContinentDaoImpl();
 		ContinentModel continentModel = continentDao.findByName(RunningGame.getInstance(), continentDto.getName());
+		if(continentModel==null) {
+			throw new RiskGameRuntimeException(String.format("Continent [%s] doesn't exist, nothing to remove!",continentDto.getName())); 
+		}
 		continentDao.getCountries(RunningGame.getInstance(), continentModel).stream().forEach(countryModel -> {
 			CountryDto countryDto = new CountryDto();
 			BeanUtils.copyProperties(countryModel, countryDto);
-			try {
-				removeCountry(countryDto);
-			} catch (Exception e) {
-				e.printStackTrace();
-			}
-		});
+			removeCountry(countryDto);
+		});	
 		continentDao.delete(RunningGame.getInstance(), continentModel);
 	}
 
@@ -54,14 +58,18 @@ public class MapService {
 	 * @param countryDto
 	 * @throws Exception
 	 */
-	public void addCountry(CountryDto countryDto) throws Exception {
-		CountryModel countryModel = objectFactory.createCountryModel();
+	public void addCountry(CountryDto countryDto){
+		CountryDaoImpl countryDaoImpl = new CountryDaoImpl();
+		CountryModel countryModel = countryDaoImpl.findByName(RunningGame.getInstance(), countryDto.getName());
+		if(countryModel!=null) {
+			throw new RiskGameRuntimeException(String.format("Country [%s] already exist, cannot be added!", countryDto.getName())); 
+		}
+		countryModel = objectFactory.createCountryModel();
 		BeanUtils.copyProperties(countryDto, countryModel);
 		ContinentDaoImpl continentDaoImpl = new ContinentDaoImpl();
 		ContinentModel continentModel = continentDaoImpl.findByName(RunningGame.getInstance(),
 				countryDto.getContenentName());
 		countryModel.setContinentId(continentModel.getId());
-		CountryDaoImpl countryDaoImpl = new CountryDaoImpl();
 		countryDaoImpl.assignID(RunningGame.getInstance(), countryModel);
 		RunningGame.getInstance().getCountries().getList().add(countryModel);
 		RunningGame.getInstance().getGraph().addVertex(countryModel.getName());
@@ -73,12 +81,15 @@ public class MapService {
 	 * @param countryDto
 	 * @throws Exception
 	 */
-	public void removeCountry(CountryDto countryDto) throws Exception {
-		BorderDaoImp borderDaoImp = new BorderDaoImp();
-		BorderModel borderModel = borderDaoImp.findByName(RunningGame.getInstance(), countryDto.getName());
-		borderDaoImp.delete(RunningGame.getInstance(), borderModel);
+	public void removeCountry(CountryDto countryDto) {
 		CountryDaoImpl countryDaoImpl = new CountryDaoImpl();
 		CountryModel countryModel = countryDaoImpl.findByName(RunningGame.getInstance(), countryDto.getName());
+		if(countryModel == null) {
+			throw new RiskGameRuntimeException(String.format("Country [%s] doesn't exist, nothing to remove!", countryDto.getName())); 
+		}
+		BorderDaoImp borderDaoImp = new BorderDaoImp();
+		BorderModel borderModel = borderDaoImp.findByName(RunningGame.getInstance(), countryDto.getName());
+		borderDaoImp.delete(RunningGame.getInstance(), borderModel);	
 		countryDaoImpl.delete(RunningGame.getInstance(), countryModel);
 		RunningGame.getInstance().getGraph().removeVertex(countryModel.getName());
 	}
@@ -87,62 +98,68 @@ public class MapService {
 	 * 
 	 * @param borderDto
 	 */
-	public void addNeighbor(BorderDto borderDto) throws Exception {
-		this.makeBorder(borderDto);
-		BorderDto borderDto2 = new BorderDto();
-		borderDto2.setCountryName(borderDto.getNeighborCountryName());
-		borderDto2.setNeighborCountryName(borderDto.getCountryName());
-		this.makeBorder(borderDto2);
+	public void addNeighbor(BorderDto borderDto){
+		CountryDaoImpl countryDaoImpl = new CountryDaoImpl();
+		CountryModel countryModel = countryDaoImpl.findByName(RunningGame.getInstance(), borderDto.getCountryName());
+		if(countryModel == null) {
+			throw new RiskGameRuntimeException(String.format("Country [%s] doesn't exist, border will be ignored", borderDto.getCountryName())); 
+		}
+		countryModel = countryDaoImpl.findByName(RunningGame.getInstance(), borderDto.getNeighborCountryName());
+		if(countryModel == null) {
+			throw new RiskGameRuntimeException(String.format("Country [%s] doesn't exist, border will be ignored", borderDto.getCountryName())); 
+		}
+		this.makeBorder(borderDto.getCountryName(), borderDto.getNeighborCountryName());
+		this.makeBorder(borderDto.getNeighborCountryName(), borderDto.getCountryName());
 		RunningGame.getInstance().getGraph().addEdge(borderDto.getCountryName(), borderDto.getNeighborCountryName());
 		
 	}
 
 	/**
 	 * 
-	 * @param borderDto
-	 * @throws Exception
+	 * @param countryName
+	 * @param neighborCountryName
 	 */
-	private void makeBorder(BorderDto borderDto) throws Exception {
+	private void makeBorder(String countryName, String neighborCountryName){
 		BorderDaoImp borderDaoImp = new BorderDaoImp();
-		BorderModel borderModel = borderDaoImp.findByName(RunningGame.getInstance(), borderDto.getCountryName());
+		BorderModel borderModel = borderDaoImp.findByName(RunningGame.getInstance(), countryName);
 		if (borderModel == null) {
 			borderModel = objectFactory.createBorderModel();
 			RunningGame.getInstance().getBorders().getList().add(borderModel);
 		}
 		CountryDaoImpl countryDaoImpl = new CountryDaoImpl();
-		CountryModel countryModel = countryDaoImpl.findByName(RunningGame.getInstance(), borderDto.getCountryName());
+		CountryModel countryModel = countryDaoImpl.findByName(RunningGame.getInstance(), countryName);
 		borderModel.setCountryId(countryModel.getId());
-		countryModel = countryDaoImpl.findByName(RunningGame.getInstance(), borderDto.getNeighborCountryName());
-		borderModel.getNeighbours().add(countryModel.getId());
+		countryModel = countryDaoImpl.findByName(RunningGame.getInstance(), neighborCountryName);
+		if(borderModel.getNeighbours().contains(Integer.valueOf(countryModel.getId()))) {
+			throw new RiskGameRuntimeException(String.format("Border [%s][%s] already exist!", countryName, neighborCountryName)); 
+		}
+		borderModel.getNeighbours().add(Integer.valueOf(countryModel.getId()));
 	}
 
 	/**
 	 * 
 	 * @param borderDto
-	 * @throws Exception
 	 */
-	public void removeNeighbor(BorderDto borderDto) throws Exception {
-		this.removeBorder(borderDto);
-		BorderDto borderDto2 = new BorderDto();
-		borderDto2.setCountryName(borderDto.getNeighborCountryName());
-		borderDto2.setNeighborCountryName(borderDto.getCountryName());
-		this.removeBorder(borderDto2);
+	public void removeNeighbor(BorderDto borderDto){
+		this.removeBorder(borderDto.getCountryName(), borderDto.getNeighborCountryName());
+		this.removeBorder(borderDto.getNeighborCountryName(), borderDto.getCountryName());
 		RunningGame.getInstance().getGraph().removeEdge(borderDto.getCountryName(), borderDto.getNeighborCountryName());
 	}
 
 	/**
 	 * 
-	 * @param borderDto
+	 * @param countryName
+	 * @param neighborCountryName
 	 */
-	private void removeBorder(BorderDto borderDto) throws Exception {
+	private void removeBorder(String countryName, String neighborCountryName){
 		BorderDaoImp borderDaoImp = new BorderDaoImp();
-		BorderModel borderModel = borderDaoImp.findByName(RunningGame.getInstance(), borderDto.getCountryName());
+		BorderModel borderModel = borderDaoImp.findByName(RunningGame.getInstance(), countryName);
 		if (borderModel == null) {
 			return;
 		}
 		CountryDaoImpl countryDaoImpl = new CountryDaoImpl();
 		CountryModel countryModel = countryDaoImpl.findByName(RunningGame.getInstance(),
-				borderDto.getNeighborCountryName());
-		borderModel.getNeighbours().remove(new Integer(countryModel.getId()));
+				neighborCountryName);
+		borderModel.getNeighbours().remove(Integer.valueOf(countryModel.getId()));
 	}
 }
